Added enums and some discussion about Drop.
The weirdest part is the idea of drops in some fields invalidating the pinning
invariant in others on the exact same structure.  To me that seems like
something that dropck is explicitly designed to avoid, but maybe some people
have a usecase for it?  My suspicion is that any problems it could cause, it
could cause equally well if the fields were split into two PinBoxes (which
implies they need a shared unsafe invariant of some sort), but I'm not certain.